Package db utils
